Feature/review edit delete

+++ docrate/src/main/java/com/team/docrate/domain/review/dto/ReviewCreateRequest.java
@@ -1,5 +1,7 @@
 package com.team.docrate.domain.review.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
@@ -16,7 +18,7 @@ public class ReviewCreateRequest {
 
     @Min(1)
     @Max(5)
-    private Double rating;
+    private BigDecimal rating;
 
     @Min(1)
     @Max(5)
